+++ PatientManager/src/main/java/com/herve/intergiciel/PatientManager/Repositories/InfoPatientRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.herve.intergiciel.PatientManager.Modeles.InfoPatient;
 
-public interface InfoPatientRepository extends JpaRepository<InfoPatient, Integer>{
+public interface InfoPatientRepository extends JpaRepository<InfoPatient, Long>{
